# Add `--precise` mode to JSON parser