… changelog/next/features/4169--json-precise.md
… libtenzir/include/tenzir/try_simdjson.hpp
… libtenzir/src/type.cpp
… tenzir/integration/data/reference/pipelines_local/test_parse_operators/step_04.ref
… tenzir/integration/data/reference/pipelines_local/test_parse_operators/step_25.ref
… tenzir/integration/data/reference/pipelines_local/test_precise_json/step_00.ref
… tenzir/integration/data/reference/pipelines_local/test_precise_json_bad_ndjson/step_00.ref
… tenzir/integration/data/reference/pipelines_local/test_precise_json_big_integer/step_00.ref
… tenzir/integration/data/reference/pipelines_local/test_precise_json_broken_input/step_00.ref
… tenzir/integration/data/reference/pipelines_local/test_precise_json_incomplete_input/step_00.ref
… tenzir/integration/data/reference/pipelines_local/test_precise_json_list_type_conflict/step_00.ref
… tenzir/integration/data/reference/pipelines_local/test_precise_json_overwrite_field/step_00.ref
… tenzir/integration/data/reference/pipelines_local/test_precise_json_raw/step_00.ref
… tenzir/integration/data/reference/pipelines_local/test_precise_json_type_conflict/step_00.ref
… tenzir/integration/tests/pipelines_local.bats
… web/docs/formats/gelf.md
… web/docs/formats/json.md