With this PR, the `json` parser gains a new `--precise` flag, which ensures that the layout of the emitted events precisely match the input. For example, it guarantees that no additional `null` fields will be added, which otherwise happens because the events are packed into the same batch.

This is especially useful with formats such as GELF, where the set fields can vary significantly between events and the user is interested in the exact fields from the input. 

Overall, the performance depends a lot on the input. When testing this with a GELF file that contains many type conflicts between IP and string, I observed a speedup from 30s to 18s (because `--precise` separates events by their types), which is a solid improvement (the majority of the time is then spent in the `data` parsers). However, when also using `--raw`, which also prevents this type conflict, `--precise` slows us down from 9s to 12s (due to being 2-pass instead of 1-pass). However, in the case of GELF, the output of `--precise` is much closer to what the user wants, so I would deem this slowdown acceptable.    

The implementation uses a custom "data signature" approach to avoid calling `type::infer` on the parsed `data`, because it turns out that the latter has unacceptable performance implications for this use-case. Also, I added a `TRY`/`TENZIR_TRY` macro with a customization point for types such as `simdjson_result`, `expected`, `optional`, `arrow::Status`, etc.